autons using new shooter
