--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/BlueBasketauto.java
@@ -1,369 +1,0 @@
-package org.firstinspires.ftc.teamcode.auto;
-
-import static org.firstinspires.ftc.teamcode.teleop.BlueTeleOp.GOAL_Y;
-import static org.firstinspires.ftc.teamcode.teleop.BlueTeleOp.GOAL_X;
-
-import com.bylazar.field.FieldManager;
-import com.bylazar.field.PanelsField;
-import com.pedropathing.follower.Follower;
-import com.pedropathing.geometry.BezierLine;
-import com.pedropathing.geometry.Pose;
-import com.pedropathing.paths.Path;
-import com.pedropathing.paths.PathChain;
-import com.pedropathing.util.Timer;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-
-import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
-import org.firstinspires.ftc.teamcode.robot.intakeLaunch;
-import org.firstinspires.ftc.teamcode.utilities.DrawingUtil;
-
-@Autonomous(name = "BLUE Basket Auto", group = "BB Auto")
-public class BlueBasketauto extends OpMode {
-
-    boolean openGateAfterPickup1 = true;
-    boolean openGateAfterPickup2 = false;
-    boolean pickupLine3 = true;
-    intakeLaunch intakeL ;
-    private Follower follower;
-    private Timer pathTimer, actionTimer, opmodeTimer;
-
-    private int pathState;
-    double launchPower = .665;
-    double transferPower = .12;
-
-    private final Pose startPose = new Pose(117, 128, Math.toRadians(45)).mirror(); // Start Pose of our robot.
-    private final Pose scorePose = new Pose(90, 90, Math.toRadians(45)).mirror();
-
-    private final Pose pickup1Pose = new Pose(100, 84, Math.toRadians(0)).mirror();
-    private final Pose pickup1PoseEnd = new Pose(128, 84, Math.toRadians(0)).mirror();// 128.5Highest (First Set) of Artifacts from the Spike Mark.
-    private final Pose gateAfterPose1 = new Pose(115,76, Math.toRadians(0)).mirror();
-    private final Pose gateAfterPose1End = new Pose(122,76, Math.toRadians(0)).mirror();
-    private final Pose scorePose2 = new Pose(96, 96, Math.toRadians(45)).mirror();// Scoring Pose of our robot. It is facing the goal at a 135 degree angle.
-    private final Pose pickup2Pose = new Pose(94, 60, Math.toRadians(0)).mirror(); // Middle (Second Set) of Artifacts from the Spike Mark.
-
-    private final Pose pickup2PoseEnd = new Pose(134, 60, Math.toRadians(0)).mirror();
-    private final Pose pickup2PoseReturn = new Pose(125,60, Math.toRadians(0)).mirror();
-//94,36
-    private final Pose pickup3Pose = new Pose(96, 36, Math.toRadians(0)).mirror(); // Lowest (Third Set) of Artifacts from the Spike Mark.
-
-    private final Pose pickup3PoseEnd = new Pose(130, 36, Math.toRadians(0)).mirror();
-    private final Pose pickup3PoseReturn = new Pose(125,36, Math.toRadians(0)).mirror();
-    private Path scorePreload;
-    private PathChain grabPickup1, grabPickup1Start, scorePickup1, grabPickup2, scorePickup2, grabPickup3, scorePickup3;
-
-    private FieldManager field;
-
-
-
-    public void buildPaths() {
-        /* This is our scorePreload path. We are using a BezierLine, which is a straight line. */
-        scorePreload = new Path(new BezierLine(startPose, scorePose));
-        scorePreload.setLinearHeadingInterpolation(startPose.getHeading(), scorePose.getHeading());
-
-//        /* This is our grabPickup1 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-
-    /* Here is an example for Constant Interpolation
-    scorePreload.setConstantInterpolation(startPose.getHeading()); */
-
-        /* This is our grabPickup1 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        grabPickup1 = follower.pathBuilder()
-                .addPath(new BezierLine(scorePose, pickup1Pose))
-                .setLinearHeadingInterpolation(scorePose.getHeading(), pickup1Pose.getHeading())
-                .addPath(new BezierLine(pickup1Pose,  pickup1PoseEnd))
-                .setLinearHeadingInterpolation(pickup1Pose.getHeading(), pickup1PoseEnd.getHeading())
-                .build();
-
-        /* This is our scorePickup1 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        if (openGateAfterPickup1) {
-            scorePickup1 = follower.pathBuilder()
-                    .addPath(new BezierLine(pickup1PoseEnd, gateAfterPose1))
-                    .setLinearHeadingInterpolation(pickup1PoseEnd.getHeading(), gateAfterPose1.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1, gateAfterPose1End))
-                    .setLinearHeadingInterpolation(gateAfterPose1.getHeading(), gateAfterPose1End.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1End, gateAfterPose1))
-                    .setLinearHeadingInterpolation(gateAfterPose1End.getHeading(), gateAfterPose1.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1, scorePose2))
-                    .setLinearHeadingInterpolation(gateAfterPose1.getHeading(), scorePose2.getHeading())
-                    .build();
-        } else {
-            scorePickup1 = follower.pathBuilder()
-                    .addPath(new BezierLine(pickup1PoseEnd, scorePose2))
-                    .setLinearHeadingInterpolation(pickup1PoseEnd.getHeading(), scorePose2.getHeading())
-                    .build();
-        }
-
-//        /* This is our grabPickup2 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        grabPickup2 = follower.pathBuilder()
-                .addPath(new BezierLine(scorePose2, pickup2Pose))
-                .setLinearHeadingInterpolation(scorePose2.getHeading(), pickup2Pose.getHeading())
-                .addPath(new BezierLine(pickup2Pose,  pickup2PoseEnd))
-                .setLinearHeadingInterpolation(pickup2Pose.getHeading(), pickup2PoseEnd.getHeading())
-                .build();
-
-        /* This is our scorePickup2 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        if (openGateAfterPickup2) {
-            scorePickup2 = follower.pathBuilder()
-                    .addPath(new BezierLine(pickup2PoseEnd, pickup2PoseReturn))
-                    .setLinearHeadingInterpolation(pickup2PoseEnd.getHeading(), pickup2PoseReturn.getHeading())
-                    .addPath(new BezierLine(pickup2PoseReturn, gateAfterPose1))
-                    .setLinearHeadingInterpolation(pickup2PoseReturn.getHeading(), gateAfterPose1.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1, gateAfterPose1End))
-                    .setLinearHeadingInterpolation(gateAfterPose1.getHeading(), gateAfterPose1End.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1End, gateAfterPose1))
-                    .setLinearHeadingInterpolation(gateAfterPose1End.getHeading(), gateAfterPose1.getHeading())
-                    .addPath(new BezierLine(gateAfterPose1, scorePose))
-                    .setLinearHeadingInterpolation(gateAfterPose1.getHeading(), scorePose.getHeading())
-                    .build();
-
-        } else {
-            scorePickup2 = follower.pathBuilder()
-                    .addPath(new BezierLine(pickup2PoseEnd, pickup2PoseReturn))
-                    .setLinearHeadingInterpolation(pickup2PoseEnd.getHeading(), pickup2PoseReturn.getHeading())
-                    .addPath(new BezierLine(pickup2PoseReturn, scorePose))
-                    .setLinearHeadingInterpolation(pickup2PoseReturn.getHeading(), scorePose.getHeading())
-                    .build();
-        }
-
-        /* This is our grabPickup3 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        grabPickup3 = follower.pathBuilder()
-                .addPath(new BezierLine(scorePose2, pickup3Pose))
-                .setLinearHeadingInterpolation(scorePose2.getHeading(), pickup3Pose.getHeading())
-                .addPath(new BezierLine(pickup3Pose,  pickup3PoseEnd))
-                .setLinearHeadingInterpolation(pickup3Pose.getHeading(), pickup3PoseEnd.getHeading())
-                .build();
-
-        /* This is our scorePickup3 PathChain. We are using a single path with a BezierLine, which is a straight line. */
-        scorePickup3 = follower.pathBuilder()
-                .addPath(new BezierLine(pickup3PoseEnd, pickup3PoseReturn))
-                .setLinearHeadingInterpolation(pickup3PoseEnd.getHeading(), pickup3PoseReturn.getHeading())
-                .addPath(new BezierLine(pickup3PoseReturn, scorePose))
-                .setLinearHeadingInterpolation(pickup3PoseReturn.getHeading(), scorePose.getHeading())
-                .build();
-    }
-
-    public void autonomousPathUpdate() {
-        switch (pathState) {
-            case 0:
-                follower.followPath(scorePreload);
-                intakeL.powerOnLauncher(launchPower);
-                setPathState(1);
-                break;
-            case 1:
-//            /* You could check for
-//            - Follower State: "if(!follower.isBusy()) {}"
-//            - Time: "if(pathTimer.getElapsedTimeSeconds() > 1) {}"
-//            - Robot Position: "if(follower.getPose().getX() > 36) {}"
-//            */
-//                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the scorePose's position */
-                if (!follower.isBusy()) {
-//                    /* Score Preload */
-                    if (!intakeLaunch.shooting) {
-                        intakeLaunch.shooting = true;
-                        intakeLaunch.runtime.reset();
-                    }
-                    intakeL.takeShot(launchPower, 2500);
-                    if (intakeLaunch.runtime.milliseconds() > 2500) {
-                        intakeLaunch.shooting = false;
-                        intakeL.stopLauncher();
-                        intakeL.runIntake(1, transferPower);
-//                    /* Since this is a pathChain, we can have Pedro hold the end point while we are grabbing the sample */
-                        follower.followPath(grabPickup1);
-                        setPathState(2);
-                    }
-                }
-                break;
-            case 2:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the pickup1Pose's position */
-                if (!follower.isBusy()) {
-                    /* Grab Sample */
-                    intakeL.runIntake(1, transferPower);
-                    /* Since this is a pathChain, we can have Pedro hold the end point while we are scoring the sample */
-                    follower.followPath(scorePickup1, true);
-                    intakeL.powerOnLauncher(.653);
-                    setPathState(3);
-                }
-                break;
-            case 3:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the scorePose's position */
-                if (!follower.isBusy()) {
-                    /* Score pick up1 */
-                    if (!intakeLaunch.shooting) {
-                        intakeLaunch.shooting = true;
-                        intakeLaunch.runtime.reset();
-                    }
-                    intakeL.takeShot(0.653, 2800);
-                    if (intakeLaunch.runtime.milliseconds() > 2800) {
-                        intakeLaunch.shooting = false;
-                        intakeL.stopIntake();
-                        intakeL.stopLauncher();
-                        intakeL.runIntake(1, transferPower);
-                        /* Since this is a pathChain, we can have Pedro hold the end point while we are grabbing the sample */
-                        follower.followPath(grabPickup2, true);
-                        setPathState(4);
-                    }
-                }
-                break;
-            case 4:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the pickup2Pose's position */
-                if (!follower.isBusy()) {
-                    /* Grab Sample */
-                    intakeL.runIntake(1, transferPower);
-                    /* Since this is a pathChain, we can have Pedro hold the end point while we are scoring the sample */
-                    follower.followPath(scorePickup2, true);
-                    intakeL.powerOnLauncher(launchPower);
-                    setPathState(5);
-                }
-                break;
-            case 5:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the scorePose's position */
-                if (!follower.isBusy()) {
-                    /* Score Sample */
-                    if (!intakeLaunch.shooting) {
-                        intakeLaunch.shooting = true;
-                        intakeLaunch.runtime.reset();
-                    }
-                    intakeL.takeShot(launchPower, 3400);
-                    if (intakeLaunch.runtime.milliseconds() > 3400) {
-                        intakeLaunch.shooting = false;
-                        intakeL.stopIntake();
-                        intakeL.stopLauncher();
-                        /* Since this is a pathChain, we can have Pedro hold the end point while we are grabbing the sample */
-                        // pick up 3rd based on flag
-                        if (pickupLine3) {
-                            intakeL.runIntake(1, transferPower);
-                            follower.followPath(grabPickup3, true);
-                        }
-                        setPathState(6);
-                    }
-                }
-                break;
-            case 6:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the pickup3Pose's position */
-                if (!follower.isBusy()) {
-                    // pick up 3rd based on flag
-                    if (pickupLine3) {
-                        /* Grab Sample */
-                        intakeL.runIntake(1, transferPower);
-                        /* Since this is a pathChain, we can have Pedro hold the end point while we are scoring the sample */
-                        follower.followPath(scorePickup3, true);
-                        intakeL.powerOnLauncher(launchPower);
-                    }
-                    setPathState(7);
-                }
-                break;
-            case 7:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the scorePose's position */
-                if (!follower.isBusy()) {
-                    // score 3rd based on flag
-                    if (pickupLine3) {
-                        /* Score Sample */
-                        if (!intakeLaunch.shooting) {
-                            intakeLaunch.shooting = true;
-                            intakeLaunch.runtime.reset();
-                        }
-                        intakeL.takeShot(launchPower, 2500);
-                        if (intakeLaunch.runtime.milliseconds() > 2500) {
-                            intakeLaunch.shooting = false;
-                            intakeL.stopIntake();
-                            intakeL.stopLauncher();
-                            follower.holdPoint(new Pose(72, follower.getPose().getX(), Math.PI/2));
-                            setPathState(8);
-                        }
-                    } else {
-                        setPathState(8);
-                    }
-                }
-                break;
-
-            case 8:
-                /* This case checks the robot's position and will wait until the robot position is close (1 inch away) from the scorePose's position */
-                if (!follower.isBusy()) {
-                    /* Set the state to a Case we won't use or define, so it just stops running an new paths */
-                    setPathState(-1);
-                }
-                break;
-        }
-    }
-
-    /**
-     * These change the states of the paths and actions. It will also reset the timers of the individual switches
-     **/
-    public void setPathState(int pState) {
-        pathState = pState;
-        pathTimer.resetTimer();
-    }
-
-    /**
-     * This is the main loop of the OpMode, it will run repeatedly after clicking "Play".
-     **/
-    @Override
-    public void loop() {
-        DrawingUtil.drawRobotOnField(field, follower.getPose().getX(), follower.getPose().getY(), follower.getPose().getHeading(), GOAL_X, GOAL_Y);
-        intakeLaunch.degrees = Math.toDegrees(follower.getHeading());
-
-        // These loop the movements of the robot, these must be called continuously in order to work
-        follower.update();
-        autonomousPathUpdate();
-//        intakeL.setTurnPosition();
-
-        // Update blackboard with current pose
-        Pose currentPose = follower.getPose();
-        blackboard.put("POSE_X", currentPose.getX());
-        blackboard.put("POSE_Y", currentPose.getY());
-        blackboard.put("POSE_HEADING", currentPose.getHeading());
-
-        // Feedback to Driver Hub for debugging
-        telemetry.addData("path state", pathState);
-        telemetry.addData("x", follower.getPose().getX());
-        telemetry.addData("y", follower.getPose().getY());
-        telemetry.addData("heading", follower.getPose().getHeading());
-        telemetry.update();
-    }
-
-    /**
-     * This method is called once at the init of the OpMode.
-     **/
-    @Override
-    public void init() {
-        field = PanelsField.INSTANCE.getField();
-        field.setOffsets(PanelsField.INSTANCE.getPresets().getPEDRO_PATHING());
-        pathTimer = new Timer();
-        opmodeTimer = new Timer();
-        opmodeTimer.resetTimer();
-
-
-        follower = Constants.createFollower(hardwareMap);
-        intakeL = new intakeLaunch(hardwareMap, telemetry);
-        intakeLaunch.initial = startPose.getHeading();
-        intakeLaunch.shooting = false;
-        buildPaths();
-        follower.setStartingPose(startPose);
-        intakeLaunch.p = 0.005;
-        intakeLaunch.f = 0.05;
-    }
-
-    /**
-     * This method is called continuously after Init while waiting for "play".
-     **/
-    @Override
-    public void init_loop() {
-    }
-
-    /**
-     * This method is called once at the start of the OpMode.
-     * It runs all the setup actions, including building paths and starting the path system
-     **/
-    @Override
-    public void start() {
-        opmodeTimer.resetTimer();
-        setPathState(0);
-    }
-
-    /**
-     * We do not use this because everything should automatically disable
-     **/
-    @Override
-    public void stop() {
-    }
-}
